In Which Tara Is Making an Attempt at Making Learning More Simple
I have the desire to make learning more simple.
I have seen that people who have knowledge of a thing do not have a good memory for how hard it was to get that knowledge.
They say to new people that 'this is simple' and make a new person have bad feelings about themselves because it is not simple to the new person.
I have a feeling that knowledge is able to be broken down into very small bits.
I have the desire to map all of those bits.
I have a feeling that learning is made much harder by giving a person 20 new bits of knowledge at a time in place of one new bit of knowledge at a time.
I have the desire to have a record of how I got every bit of knowledge.
I have the desire to make the discovery of how to give a person one new bit of knowledge at a time in place of 20.
I do not have a good idea of how I will do this.
So I am making this record so that I am able to see how I did it.