In Which Tara says "What is Done?" "I need a test to see when I am 'done'"
I does not like not having an idea of if I am done with a bit of work.
I have a theory that in order to get knowledge, I need be able to test if she has the bit of knowledge that I was making the attempt to get.
If I have a way to test if she has gotten that knowledge, I will be able to say "I am done getting this bit of knowledge, or I am not done getting this bit of knowledge."
Being able to say I is done getting a bit of knowledge will let me stop working on that bit and start working on getting a new bit.

So in order to get a bit of knowledge, I will
1. make the attempt to make a test for if I have gotten the knowledge
2. make an attempt at learning the knowledge.
3. take the test to see if I have have gotten the knowledge