In Which Tara makes her first attempt to make learning more simple
My theory is this:
When people make the attempt to give a person one new bit of knowledge, they give 20 new bits of knowledge.
For example, I may make the attempt to say how to make a meal to a person.  I make an error, because the person I am saying this to does not have knowledge of my language, English. For them to get knowledge of how to make a meal from me, they have to get knowledge of English.  To get the knowledge of how to make a meal from me, they have to get 10,000 bits of knowledge of how to say things in English.  It is much more simple for that person to go to a person who is able to say how to make a meal in their language.  The only new things they have a need to be learning are the bits about how to make a meal.

I have the desire to have people learning 1 bit of knowledge at a time in place of 10,000.  I am starting like this:

I only use 850 common English words.  I got these words from http://ogden.basic-english.org/
It is possible to use only these 850 words to say almost any idea.

This is more simple because people only need to have knowledge of these 850 words.

When I use a word that is not on the list, that word is a new bit of knowledge.  It is a new idea.
Because it is a new idea, I put it at the top of the page.
So every new idea in a page is at the top, so people are able to the things they will be learning.

With those ideas, I say what the idea is, and I put a picture of the idea.

This helps me, because if I am able to say to myself what the idea is, that is a test that I have that bit of knowledge.
If I am able to say what a thing is like, that is a test that I have that bit of knowledge
If I am able to put a picture of that idea, that is a test I have that bit of knowledge

If I have those three things, I may have a good chance of being able to give that knowledge to a new person.

Because all new ideas are at the start, I am able to be certain that I am giving the bits of knowledge that I am making the attempt to give, and not giving to much, or even worse, not giving an important bit of knowledge.

I am going to make the attempt to use this system and see if it helps me note down all of the bits of knowledge I am learning.