In which Tara makes a system of noting ideas
New Word,what it is,what it is like,picture
csv,is a way of making a list of ideas,a shelf of books,WhatIsDoneMemes/csvIsLikeBookshelf.gif

At the top of the page are the new ideas.
So far, I have made the decision that for me to test if I have knowledge of an idea,
1. I need a word for it
2. I need to be able to say what it is
4. I need to be able to say what it is like
5. I need to be able to have a picture of it.

I am using a way of writing that is named csv.

csv is a way of making a list of ideas
you put a ',' between the ideas, so that you have a list
At the top, you put a list of the kind of ideas.

So,

New Word,what it is,what it is like,picture

are the names of the kinds of ideas that go in the list, and

kung fu,is the art of kicking people if they are making the attempt to kick you,it is like being like water,WhatIsDoneMemes/KungFu.gif

are ideas in the list.

Put them together, and what you see is this

New Word,what it is,what it is like,picture
kung fu,is the art of kicking people if they are making the attempt to kick you,it is like being like water,WhatIsDoneMemes/KungFu.gif