In which Tara gives a list of all the words she is starting with
These are all of the words I am starting with.  I am using the 850 words from http://ogden.basic-english.org/
What is hard is that the 850 words have different forms.  When I put the forms in the list, there are over 3000.
These lists are full of errors, and I have the desire to make better lists, but at this time these lists are
good enough that I am able to use them to help me not use complex words.