In which Tara says how she keeps her words so simple
I made a thing named a Syntax Higlighter. It makes the words that I am letting myself use brown, and it makes the words that are not on that list black. That way I am able to see if I am using the right language.

This syntax Highlighter only works with Sublime, and it is put together in a very bad way.  It is full of errors.  It is good enough that I am able to use it to keep my words simple. I am not going to talk about how to put it on a different person's computer for now, because I have the desire for it to not be full of errors when I give it to other people.