In which Tara makes a machine that gives her a list of all words that are not in Ogden's Basic 850
I made this thing. The good thing about it is that it works. The bad thing about it is that it is a bad bit of writing. I will need to do a great number of things to it if I were to give it to a different person. I have knowledge that the word list I use has errors in it.

I may or may not make this machine better, because I am not certain if it is going to be useful. If I do need it later, I am going to do it again the right way.