In which Tara Tells git to ignore .DS_Store files
New Word,what it is,what it is like,picture
git repository,is a place to store all of the writings you have a desire to keep safe,is like a library,no picture

I do not have an idea of what a .DS_Store file is. I do have the opinion
that I do not need to have them in my in my git repository.