In Which Tara gets .DS_Store out of her git repository
word,what it is,what it is like, picture
computer,is a machine that will do the things I say if I say it in the right language,is like a dog
folder,is a place to put a number of different writings, is like a box to put things in,no picture
.DS_Store,is a writing that says to my computer when you are in this folder make the things in this folder look a certain way,is like the colours of paint on a wall,no picture

I do not have the desire to have .DS_Store in my git repository. So I am taking it out. I do not have the need for people to see the .DS_Store for the same reason I do not have a need for people to  see what colours I have painted the walls of my room.