In Which Tara Moves a bit of writing into the folder of the day it was Made
This is to keep things in order. I have the idea that it is more simple to go back in time and see what I did on a day. This helps me keep myself in order because I have noted that if I see a more than three things in a place, it is hard for me to make a decision about what to do. Putting the days work in a day folder will help me keep my day's work from being put in a great number of different places.