In Which Tara Makes a List of Things She Does Not have Time To Be Writing About
I put down three things that people who are writing programming books do not ever make directions for.
Two of the things are things you need to be able to use to be a good programmer who makes good programs. It is more simple to
do these things from the start, because it is hard to make yourself do these thing later.

I do not have the desire to be writing about these things because I have the desire to be writing about programming.  I like writing more about
programming than I like writing about using git or the command line.  Most programmers are like me, which is why there is a very small number of
writings that say how to use these in simple language.