In which Tara makes the decision to start with Javascript as a first programming language
installation,is to put a program on your computer,is like putting a new washing machine in a house
browser,is a program that lets you see things on the internet,is like a window to a different place
internet,is the greatest store of knowledge in history,is like a library

I made the decision to use javascript because javascript is on your computer in a browser. You do not have to do complex things to get started.