In which Tara makes a page with nothing on it
word,what it is,what it is like
webpage,is a bit of writing on the internet,is like a book
Google Chrome,is a program for looking at webpages on the internet,is like a window you are able to see a book through

I made this page using only Google Chrome.
I now see that it is possible to make webpages using only Google Chrome.