In which tara says how to make a picture so that you are not able to see it
I have the opinion that being able to make parts of a webpage not be in view is a good thing to be able to do.
It lets you make parts of a webpage that are not simple go away.
This lets you put your attention on the part you have the desire to be learning.