In which Tara says how to make a picture go away
word,what it is,what it is like
gif,a picture that is in motion,a motion picture
LICECap,a program that makes gifs, a machine for taking pictures

I made a page that says to people how to make pictures go out of view.

My brother said to me that "LICEcap is a very good program." It is in fact a very good program. It is used to make gifs.
It is more simple to let a person see how to do a thing than to say to them how to do it step by step. I was able to a great number of pictures into two small motion pictures.
This makes it much more simple to let people see what to do.

There are a great number of pages with too much on them. This is the first step in letting people that take their attention away from what is important.