In which Tara changes the page to use a different way of making a rabbit not be in view
word,what it is,what it is like
CSS,is a way of making changes to how things look on a page,is like paint

I made the decision to use a thing named visibility:hidden to make my rabbit go away in place of display:none.

In this page I am giving people the knowledge of how to make it so that you do not have to see things you do not have a desire to see.  I do this so that I am able to put my attention to the more important things and not have my attention go every which way.

I gave people one line of CSS that they are able to use to make a picture go away.  In this page, I use visibilty hidden. This is like taking a picture off a wall.  The picture is gone, but the space is still there.

Before, I was using display:none.  This is like getting a bottle from an icebox in a store. a new bottle goes right in the place that you took the old one from, and you would never have any idea that the old bottle was there in the first place.