How to make a rabbit picture so you are not able to see it
I have the opinion that this is a better way to start learning to program than
"hello world." It lets you do a thing that you see.

I am angry when I see a page with a great number of things that are not necessary.  This trick lets you make the things you do not have a desire to see go away.