In which Tara makes it so that you are able to save this page for viewing when you do not have a connection
I was having trouble with the object tag. When I went to save the page like give directions to do on this page,
the movies would go away when I made my net connection go away. I looked at the code, and the in the data attribute it was
giving me an absolute path. Chrome  had the thought that the object tag was making a connection to a different page on the net. It did not
have the opinion that the things in the tag were important parts of the page. It is probable Chrome had the thought that these were advertisements.
So it did not save them.

When Chrome saves a website, there are many things that come with it. It is like a used car. When you get the car, the person who giving the car to you may give you a great number of things to go with the car.

When you use the object tag, it is like putting another car in your car- a motorcycle.  The person who is giving you the first car does not desire to give you the motorcycle as well.  So he takes the motorcycle out and lets you go away with the car. That is why Chrome was not giving me the things in the object tags. It thought that, like the motorcyle, those things were not mine to take.

I was using the object tag to view my pictures. All of the cars are mine.

I made the discovery that when the pictures are in other places on the page, when Chrome saves the objects, it saves the data links as a relative path.  Since Chrome sees the pictures in other places on the page, it now has the opinion that those pictures are mine and I can take them when I save the page.